Added mime-detecive, adding services
