+++ Sulfur/Services/UrlHeaderActions/IUrlActionService.cs
@@ -9,7 +9,6 @@ namespace Sulfur.Services.UrlHeaderActions
     public interface IUrlActionService
     {
         bool SuccessfulMatchOnHeaderToken(String headerToken);
-        bool ValidUrl(String url);
 
         GuidResult GenerateGuidPayload();
     }
