+++ Sulfur/Services/UrlHeaderActions/UrlActionService.cs
@@ -26,33 +26,6 @@ namespace Sulfur.Services.UrlHeaderActions
             return guidCode;
         }
 
-        //Returns whether the url is valid
-        public bool ValidUrl(string url)
-        {
-            HttpWebResponse response = null;
-            var request = (HttpWebRequest)WebRequest.Create(url);
-            request.Method = "HEAD";
 
-
-            try
-            {
-                response = (HttpWebResponse)request.GetResponse();
-            }
-            catch (WebException ex)
-            {
-                /* A WebException will be thrown if the status of the response is not `200 OK` */
-                return false;
-            }
-            finally
-            {
-                // Don't forget to close your response.
-                if (response != null)
-                {
-                    response.Close();
-                }
-            }
-
-            return true;
-        }
     }
 }
